+++ packages/agtree/CHANGELOG.md
@@ -7,6 +7,14 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
+## Unreleased <!-- release/browser-extension-v5.4 -->
+
+## Added
+
+- Path-in-domain syntax support for cosmetic rules [#169].
+
+[#169]: https://github.com/AdguardTeam/tsurlfilter/issues/169
+
 ## Unreleased <!-- release/browser-extension-v5.3 -->
 
 ### Added